Add a bunch of libraries, programatically add flippers
This was a little hairy. Needed to fall back on some backbone. This gives
classic event handling between objects. I am using the spindle as a service
object, and adding the event handling to that for corrodination. Seems Angular
pratices would prefer to go through the scope, but I didn't see this right off,
but sould work well. We'll see how the code shapes up.

Nice to have the fall back, and with the added libraries, nothing conflicts, and
infact, angular use jQuery instead of jqLite in this case.

Keeping all of this in mind.

Would be nice to construct the directive, or get an instance with out the
compile, but seems as that is the most direct route. It would seem, the repeater
should be used, but waiting for the bind to take place seems slow. Trying to see
if an imperitive style of coding would work better for the spindle, where we add
and remove from the ends to get a rotation effect.

Just noting: The communication is between the service and the directive, which
we don't neccearily have access too. Nor the the directive have an object.
Something to keep in mind, as learning Angular. Would be nice to simply grab the
element, and load against it. That is probably the ideal road, as opposed to
using the Backbone.Events like I am doing.

The controller maybe the place to do this, and then perhaps event handling
through the scope will work, seeing as the directive and the controller
basically share a scope.