Implement left and right move for spindle:
Regardless of correct implementation, this version works much more speedily that
the ng-repeater binding version, and seems more correct in a sense.

Leave the door open for the reading from the inbetween section to have a larger
implied spindle with more entries, which perhaps still only having a certian
number of flippers.

We'll have to see...